move classes to dedicated files [skip ci]
