+++ src/webview/android/kotlin/RustWebChromeClient.kt
@@ -493,179 +493,3 @@ class RustWebChromeClient(appActivity: AppCompatActivity) : WebChromeClient() {
     return File.createTempFile(imageFileName, ".jpg", storageDir)
   }
 }
-
-object PermissionHelper {
-  /**
-   * Checks if a list of given permissions are all granted by the user
-   *
-   * @since 3.0.0
-   * @param permissions Permissions to check.
-   * @return True if all permissions are granted, false if at least one is not.
-   */
-  fun hasPermissions(context: Context?, permissions: Array<String>): Boolean {
-    for (perm in permissions) {
-      if (ActivityCompat.checkSelfPermission(
-          context!!,
-          perm
-        ) != PackageManager.PERMISSION_GRANTED
-      ) {
-        return false
-      }
-    }
-    return true
-  }
-
-  /**
-   * Check whether the given permission has been defined in the AndroidManifest.xml
-   *
-   * @since 3.0.0
-   * @param permission A permission to check.
-   * @return True if the permission has been defined in the Manifest, false if not.
-   */
-  fun hasDefinedPermission(context: Context, permission: String): Boolean {
-    var hasPermission = false
-    val requestedPermissions = getManifestPermissions(context)
-    if (requestedPermissions != null && requestedPermissions.isNotEmpty()) {
-      val requestedPermissionsList = listOf(*requestedPermissions)
-      val requestedPermissionsArrayList = ArrayList(requestedPermissionsList)
-      if (requestedPermissionsArrayList.contains(permission)) {
-        hasPermission = true
-      }
-    }
-    return hasPermission
-  }
-
-  /**
-   * Check whether all of the given permissions have been defined in the AndroidManifest.xml
-   * @param context the app context
-   * @param permissions a list of permissions
-   * @return true only if all permissions are defined in the AndroidManifest.xml
-   */
-  fun hasDefinedPermissions(context: Context, permissions: Array<String>): Boolean {
-    for (permission in permissions) {
-      if (!hasDefinedPermission(context, permission)) {
-        return false
-      }
-    }
-    return true
-  }
-
-  /**
-   * Get the permissions defined in AndroidManifest.xml
-   *
-   * @since 3.0.0
-   * @return The permissions defined in AndroidManifest.xml
-   */
-  private fun getManifestPermissions(context: Context): Array<String>? {
-    var requestedPermissions: Array<String>? = null
-    try {
-      val pm = context.packageManager
-      val packageInfo = pm.getPackageInfo(context.packageName, PackageManager.GET_PERMISSIONS)
-      if (packageInfo != null) {
-        requestedPermissions = packageInfo.requestedPermissions
-      }
-    } catch (ex: Exception) {
-    }
-    return requestedPermissions
-  }
-
-  /**
-   * Given a list of permissions, return a new list with the ones not present in AndroidManifest.xml
-   *
-   * @since 3.0.0
-   * @param neededPermissions The permissions needed.
-   * @return The permissions not present in AndroidManifest.xml
-   */
-  fun getUndefinedPermissions(context: Context, neededPermissions: Array<String?>): Array<String?> {
-    val undefinedPermissions = ArrayList<String?>()
-    val requestedPermissions = getManifestPermissions(context)
-    if (requestedPermissions != null && requestedPermissions.isNotEmpty()) {
-      val requestedPermissionsList = listOf(*requestedPermissions)
-      val requestedPermissionsArrayList = ArrayList(requestedPermissionsList)
-      for (permission in neededPermissions) {
-        if (!requestedPermissionsArrayList.contains(permission)) {
-          undefinedPermissions.add(permission)
-        }
-      }
-      var undefinedPermissionArray = arrayOfNulls<String>(undefinedPermissions.size)
-      undefinedPermissionArray = undefinedPermissions.toArray(undefinedPermissionArray)
-      return undefinedPermissionArray
-    }
-    return neededPermissions
-  }
-}
-
-class Logger {
-  companion object {
-    private const val LOG_TAG_CORE = "Tauri"
-
-    fun tags(vararg subtags: String): String {
-      return if (subtags.isNotEmpty()) {
-        LOG_TAG_CORE + "/" + TextUtils.join("/", subtags)
-      } else LOG_TAG_CORE
-    }
-
-    fun verbose(message: String) {
-      verbose(LOG_TAG_CORE, message)
-    }
-
-    private fun verbose(tag: String, message: String) {
-      if (!shouldLog()) {
-        return
-      }
-      Log.v(tag, message)
-    }
-
-    fun debug(message: String) {
-      debug(LOG_TAG_CORE, message)
-    }
-
-    fun debug(tag: String, message: String) {
-      if (!shouldLog()) {
-        return
-      }
-      Log.d(tag, message)
-    }
-
-    fun info(message: String) {
-      info(LOG_TAG_CORE, message)
-    }
-
-    fun info(tag: String, message: String) {
-      if (!shouldLog()) {
-        return
-      }
-      Log.i(tag, message)
-    }
-
-    fun warn(message: String) {
-      warn(LOG_TAG_CORE, message)
-    }
-
-    fun warn(tag: String, message: String) {
-      if (!shouldLog()) {
-        return
-      }
-      Log.w(tag, message)
-    }
-
-    fun error(message: String) {
-      error(LOG_TAG_CORE, message, null)
-    }
-
-    fun error(message: String, e: Throwable?) {
-      error(LOG_TAG_CORE, message, e)
-    }
-
-    fun error(tag: String, message: String, e: Throwable?) {
-      if (!shouldLog()) {
-        return
-      }
-      Log.e(tag, message, e)
-    }
-
-    private fun shouldLog(): Boolean {
-      return true
-    }
-  }
-}
